Refactored pivot and fact table code to different codebase(file) from topology set
